ai-video-creator-python/requirements.txt
# AI Video Creator - Python Worker (Lean)
# Only video processing dependencies — no web framework, no storage, no crypto

# ── TTS (Text to Speech) ──────────────────────────────────────
edge-tts>=6.1.9            # Microsoft Edge TTS, FREE, supports Chinese

# ── Video Processing ──────────────────────────────────────────
moviepy==1.0.3             # Video composition
Pillow>=9.0.0              # Image processing
numpy>=1.24.0

# ── Subtitle ──────────────────────────────────────────────────
pysrt>=1.1.2               # SRT subtitle file generation

# ── HTTP Client (Pexels API) ─────────────────────────────────
httpx>=0.27.0              # HTTP client for media fetching

# ── Configuration ────────────────────────────────────────────
pydantic>=2.6.0            # Data validation
pydantic-settings>=2.2.0   # Settings from env

# ── RocketMQ (Java bridge) ───────────────────────────────────
rocketmq-client-python>=2.0.0  # MQ consumer/producer
